Added pre-validation of mandatory attributes
