+++ backend/app/src/main/java/com/moh/phlat/backend/service/ControlService.java
@@ -1,0 +1,5 @@
+package com.moh.phlat.backend.service;
+
+public interface ControlService {
+    public void setControlStatus(Long controlId, String statusCode, String authenticatedUserId);
+}
