+++ backend/app/src/main/resources/messages.properties
@@ -1,2 +1,3 @@
 columns.displayPreference.not.found=Columns display preference not found
 columns.displayPreference.invalidViewName=Allowed values for view names are {0},{1},{2},{3}
+processData.mandatoryColumnDisplayName=Error Code: 100 - {0} is mandatory.
